2000-1.67

+++ include/function.h
@@ -3,14 +3,14 @@
 
 typedef struct Matrix
 {
-    int row;
-    int column;
+    size_t row;
+    size_t column;
     float *data;
 } Matrix;
 
-struct Matrix *createZero(int row, int column);
-struct Matrix *createSpe(int row, int column, const float *data);
-struct Matrix *createRam(int row, int column, int range);
+struct Matrix *createZero(size_t row, size_t column);
+struct Matrix *createSpe(size_t row, size_t column, const float *data);
+struct Matrix *createRam(size_t row, size_t column, int range);
 void deleteMatrix(struct Matrix **matrix); //删除矩阵
 
 void matmul_plain(const struct Matrix *matrix1, const struct Matrix *matrix2, struct Matrix *matrix3); //矩阵相乘
